refactor: update auth pages
