+++ src/app/(auth)/login/page.tsx
@@ -1,7 +1,5 @@
 'use client'
 
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Label } from '@radix-ui/react-label'
 import { CircleAlert } from 'lucide-react'
@@ -9,6 +7,9 @@ import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { z } from 'zod'
+
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
 
 const schema = z.object({
   email: z
@@ -60,83 +61,92 @@ export default function Login() {
   }
 
   return (
-    <main className="flex min-h-screen items-center justify-center">
+    <main className="flex min-h-screen items-center justify-center bg-slate-50">
       <div className="mr-16 h-[570px] w-60 bg-slate-600" />
 
-      <div className="flex gap-8">
-        <div className="flex min-h-[610px] w-[420px] flex-col rounded-md border-2 px-8 pt-9 pb-8">
-          <form onSubmit={handleSubmit(onSubmit)} className="space-y-28">
+      <div className="flex min-h-[610px] w-[420px] flex-col rounded-md border px-8 py-9">
+        <form
+          onSubmit={handleSubmit(onSubmit)}
+          className="flex flex-grow flex-col justify-between"
+        >
+          <div>
             <div>
-              <h1 className="font-semibold text-[32px] text-slate-900">
+              <h1 className="font-semibold text-[32px] text-slate-900 leading-none">
                 Login
               </h1>
 
-              <p className="pt-3 text-base">Que bom receber você novamente!</p>
+              <p className="pt-3 text-base text-slate-500 leading-tight">
+                Que bom receber você novamente!
+              </p>
+            </div>
 
-              <div className="pt-10">
-                <Label className="text-[14px] text-slate-900">
+            <div className="flex w-full flex-col items-center gap-5 pt-8">
+              <div className="w-full">
+                <Label className="text-slate-900 text-sm leading-none">
                   E-mail Institucional
                 </Label>
 
                 <Input
-                  className={`mt-3 w-[356px] border-2 text-base placeholder-slate-700 focus:border-none focus:outline-none focus:ring-0 focus:ring-slate-500 ${
-                    loginFailed || errors.email ? 'border-red-800' : ''
-                  }`}
+                  className="mt-2"
+                  variant={errors.email ? 'error' : 'default'}
                   type="text"
-                  placeholder="Insira seu e-mail"
+                  placeholder="agomes@alu.ufc.br"
                   {...register('email')}
                   onBlur={() => trigger('email')}
                 />
               </div>
 
-              <div className="pt-6">
-                <Label className="text-[14px] text-slate-900">Senha</Label>
+              <div className="w-full">
+                <Label className="text-slate-900 text-sm leading-none">
+                  Senha
+                </Label>
 
                 <Input
-                  className={`mt-3 w-[356px] border-2 text-base placeholder-slate-700 focus:border-none focus:outline-none focus:ring-0 focus:ring-slate-500 ${
-                    loginFailed || errors.password ? 'border-red-800' : ''
-                  }`}
+                  className="mt-2"
+                  variant={errors.password ? 'error' : 'default'}
                   type="password"
-                  placeholder="Insira sua senha"
+                  placeholder="••••••"
                   {...register('password')}
                   onBlur={() => trigger('password')}
                 />
-                <p className="pt-3 text-slate-600 text-xs">
+
+                <p className="pt-3 font-medium text-slate-600 text-xs">
                   Pelo menos 6 caracteres
                 </p>
               </div>
-
-              {/* Mensagem de erro genérica */}
-              {loginFailed && (
-                <div className="flex items-center gap-3 pt-3">
-                  <CircleAlert className="h-4 w-4 text-red-800" />
-                  <p className=" text-red-800 text-sm">Credenciais inválidas</p>
-                </div>
-              )}
             </div>
 
-            <div>
-              <Button
-                type="submit"
-                disabled={!isValid}
-                className={`w-[356px] rounded-md bg-slate-700 py-2 text-slate-100 ${
-                  isValid ? 'cursor-pointer' : 'cursor-not-allowed opacity-50'
-                }`}
-              >
-                Login
-              </Button>
-              <Button
-                type="button"
-                onClick={() => router.push('/register')}
-                className="mt-4 w-[356px] rounded-md bg-slate-200 py-2 text-slate-600"
-              >
-                Criar uma Conta
-              </Button>
-            </div>
-          </form>
-        </div>
-        <div className="mr-16 h-[570px] w-60 bg-slate-600" />
+            {/* Mensagem de erro genérica */}
+            {loginFailed && (
+              <div className="flex items-center gap-3 pt-6">
+                <CircleAlert className="h-4 w-4 text-red-800" />
+                <p className=" text-red-800 text-sm">Credenciais inválidas</p>
+              </div>
+            )}
+          </div>
+
+          <div>
+            <Button
+              className="w-full"
+              variant="dark"
+              type="submit"
+              disabled={!isValid}
+            >
+              Login
+            </Button>
+
+            <Button
+              type="button"
+              onClick={() => router.push('/register')}
+              className="mt-4 w-full rounded-md bg-slate-200 py-2 text-slate-600"
+            >
+              Criar uma Conta
+            </Button>
+          </div>
+        </form>
       </div>
+
+      <div className="ml-16 h-[570px] w-60 bg-slate-600" />
     </main>
   )
 }
