+++ src/components/ui/button.tsx
@@ -10,6 +10,7 @@ const buttonVariants = cva(
     variants: {
       variant: {
         default: 'bg-slate-200 text-slate-600 hover:bg-slate-300',
+        dark: 'bg-slate-700 text-slate-100 hover:bg-slate-900',
         outline:
           'border border-slate-200 bg-white hover:bg-slate-100 hover:text-slate-900 dark:border-slate-800 dark:bg-slate-950 dark:hover:bg-slate-800 dark:hover:text-slate-50',
         secondary:
@@ -21,7 +22,7 @@ const buttonVariants = cva(
       size: {
         default: 'px-4 py-2 text-base',
         sm: 'rounded-md px-3 py-2 text-sm',
-        lg: 'h-11 rounded-md px-8',
+        lg: 'px-4 py-3',
         icon: 'h-10 w-10',
       },
     },
