+++ src/components/ui/input.tsx
@@ -3,25 +3,28 @@ import * as React from 'react'
 
 import { cn } from '@/lib/utils'
 
-const inputVariants = cva('w-full rounded-[6px] border transition-colors', {
-  variants: {
-    variant: {
-      default:
-        'border-slate-300 bg-slate-100 text-slate-900 placeholder-slate-500 hover:bg-slate-200',
-      error:
-        'border-red-800 bg-slate-100 text-slate-900 placeholder-slate-500 hover:bg-slate-200',
+const inputVariants = cva(
+  'w-full rounded-[6px] border outline-none transition-colors focus:outline-none',
+  {
+    variants: {
+      variant: {
+        default:
+          'border-slate-300 bg-slate-100 text-slate-900 placeholder-slate-500 hover:bg-slate-200',
+        error:
+          'border-red-800 bg-slate-100 text-slate-900 placeholder-slate-500 hover:bg-slate-200',
+      },
+      inputSize: {
+        default: 'p-3',
+        md: 'px-3 py-2',
+        sm: 'px-2.5 py-1.5',
+      },
     },
-    inputSize: {
-      default: 'p-3',
-      md: 'px-3 py-2',
-      sm: 'px-2.5 py-1.5',
+    defaultVariants: {
+      variant: 'default',
+      inputSize: 'default',
     },
   },
-  defaultVariants: {
-    variant: 'default',
-    inputSize: 'default',
-  },
-})
+)
 
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement>,
